backend/app/requirements.txt
CacheControl==0.12.11
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==39.0.2
decorator==5.1.1
dnspython==2.3.0
email-validator==1.3.1
firebase-admin==6.1.0
Flask==2.2.3
Flask-WTF==1.1.1
google-api-core==2.11.0
google-api-python-client==2.80.0
google-auth==2.16.2
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.10.0
google-cloud-storage==2.7.0
google-cloud-tasks==2.13.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
grpc-google-iam-v1==0.12.6
grpcio==1.51.3
grpcio-status==1.51.3
gunicorn==20.1.0
httplib2==0.21.0
idna==3.4
importlib-metadata==6.0.0
infinity==1.5
intervals==0.9.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
msgpack==1.0.5
phonenumbers==8.13.7
proto-plus==1.22.2
protobuf==4.22.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.6.0
pyparsing==3.0.9
python-dotenv==1.0.0
requests==2.28.2
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.5.post1
SQLAlchemy-Utils==0.40.0
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.14
validators==0.20.0
Werkzeug==2.2.3
WTForms==3.0.1
WTForms-Components==0.10.5
zipp==3.15.0
